[Refactor] Refactor Update functions
