+++ openstack/cce/v3/clusters/results.go
@@ -62,9 +62,3 @@ func (r GetCertResult) ExtractMap() (map[string]interface{}, error) {
 	err := r.ExtractInto(&s)
 	return s, err
 }
-
-// UpdateIpResult represents the result of an update operation. Call its Extract
-// method to interpret it as a Cluster.
-type UpdateIpResult struct {
-	golangsdk.ErrResult
-}
