--- openstack/cce/v3/clusters/urls.go
@@ -1,17 +1,0 @@
-package clusters
-
-import golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
-
-const (
-	rootPath     = "clusters"
-	certPath     = "clustercert"
-	masterIpPath = "mastereip"
-)
-
-func certificateURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(rootPath, id, certPath)
-}
-
-func masterIpURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(rootPath, id, masterIpPath)
-}
